Remove PostUninstaller and move it's logic to Deletion, update tests
